pyproject.toml
[tool.poetry]
name = "scraping-and-eda-marks"
version = "0.1.0"
description = ""
authors = ["Fernando Martinez Gomez <fermago2005@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
bs4 = "^0.0.2"
selenium = "^4.27.1"
python-dotenv = "^1.0.1"
matplotlib = "^3.10.0"
scipy = "^1.14.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
